Finished text score effect script
I added some more comments to the text score effect script and finished testing.

This script uses the Ballmovement script to detect when a point is scored. When this happens it enlarges the score number of the player that scored and fades it away. After which it returns the score number to original size and resets it opacity.